backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15,<4.0
Pillow>=11.0.0
psycopg[binary]>=3.2.0
celery>=5.4,<6.0
redis>=5.0,<6.0
django-cors-headers>=4.3,<5.0
drf-spectacular>=0.27,<0.28
python-dotenv>=1.0,<2.0
openai>=1.0,<2.0
cryptography>=43.0,<44.0
astral>=3.2,<4.0
djangorestframework-simplejwt[crypto]>=5.3.0,<6.0

